Restructure developer tokens feature into plugin based architecture
