--- Application/Features/DeveloperTokens/Services/DeveloperTokenManager.cs
@@ -1,59 +1,0 @@
-﻿using Application.Features.DeveloperTokens.DTO;
-using Application.Features.DeveloperTokens.Interfaces;
-using Domain.Features.DeveloperTokens;
-using Domain.Features.DeveloperTokens.Repositories;
-using Domain.Features.DeveloperTokens.ValueObject;
-
-namespace Application.Features.DeveloperTokens.Services;
-
-/// <summary>
-/// Manager responsible for handling developer token operations.
-/// </summary>
-/// <remarks>
-/// <list type="bullet">
-/// <item>Provides creation, deletion, and retrieval of developer tokens.</item>
-/// <item>Delegates JWT generation to <see cref="IDeveloperTokenService"/>.</item>
-/// <item>Persists tokens using <see cref="IDeveloperTokenRepository"/>.</item>
-/// </list>
-/// </remarks>
-public class DeveloperTokenManager(
-    IDeveloperTokenService tokenService,
-    IDeveloperTokenRepository repository
-    ) : IDeveloperTokenManager
-{
-    /// <inheritdoc />
-    public async Task<DeveloperTokenCreated> CreateAsync(Guid developerId, string name, IEnumerable<string> scopes, TimeSpan? lifetime = null,
-        CancellationToken ct = default)
-    {
-        Console.WriteLine("Debug DeveloperTokenManager");
-        var token = DeveloperToken.Create(
-            developerId,
-            name,
-            scopes.Select(s => (TokenScope)s),
-            lifetime
-        );
-
-        //await repository.SaveAsync(token, ct);
-        var tokenPair = await tokenService.GenerateToken(token);
-        
-        return new DeveloperTokenCreated
-        {
-            Token = token, 
-            ShortKey = tokenPair.ShortKey,
-            Jwt = tokenPair.Jwt
-        };
-    }
-    
-    /// <inheritdoc />
-    public async Task DeleteAsync(Guid tokenId, CancellationToken ct = default) =>
-        await repository.DeleteAsync(tokenId, ct);
-    
-    /// <inheritdoc />
-    public async Task<IReadOnlyList<DeveloperToken>> GetByDeveloperAsync(Guid developerId, CancellationToken ct = default) =>
-        await repository.GetByDeveloperIdAsync(developerId, ct);
-    
-    /// <inheritdoc />
-    public async Task<DeveloperToken?> GetByIdAsync(Guid tokenId, CancellationToken ct = default) =>
-        await repository.GetByIdAsync(tokenId, ct);
-    
-}
