+++ src/Plugins/Solutions/DevTokens/src/Repositories/IDeveloperTokenRepository.cs
@@ -1,4 +1,4 @@
-﻿namespace Domain.Features.DeveloperTokens.Repositories;
+﻿namespace DevTokens.Repositories;
 
 /// <summary>
 /// Repository interface for managing <see cref="DeveloperToken"/> entities.
